+++ Elin/LayerShippingResult.cs
@@ -109,7 +109,7 @@ public class LayerShippingResult : ELayer
 		}
 		FactionBranch branch = (ELayer.game.spatials.Find(result.uidZone) ?? ELayer.pc.homeZone).branch;
 		float num = (float)update / (float)maxUpdate;
-		int income = result.GetIncome();
+		long income = result.GetIncome();
 		income = (int)((float)income * num);
 		long num2 = result.total + income;
 		textIncome.text = Lang._currency(income, showUnit: true);
